requirements.txt
scipy==1.2.1
numpy==1.19.1
matplotlib==3.3.1
pillow==7.2.0
pandas==1.1.1
gdal==2.4.2
requests==2.24.0
skgeom==0.1.2
shapely==1.6.4
lxml==4.5.1
